Migracion GP/pyproject.toml
[project]
name = "migracion-gp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "neo4j>=6.0.3",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlglot>=28.3.0",
    "sqlparse>=0.5.4",
]
